app/requirements.txt
aiofiles
cffi
cryptography
docutils
dotty-dict
gitdb
httptools
jwt
multidict
pycparser
PyJWT
sanic
sanic-routing
ujson
uvloop
websockets
google-cloud-storage
abifsm @ git+https://github.com/voteagora/abifsm@master
PyYAML
argh
web3
psycopg2-binary
websocket-client
sanic-ext